--- Tests/ConnectionHandler/Tests_ConnectionHandler.h
@@ -1,11 +1,0 @@
-//
-// Created by ogrebatel on 03.12.22.
-//
-
-#ifndef MOND_DB_TESTS_CONNECTIONHANDLER_H
-#define MOND_DB_TESTS_CONNECTIONHANDLER_H
-
-#include "IConnectionHandler.h"
-#include "ConnectionHandler.h"
-
-#endif //MOND_DB_TESTS_CONNECTIONHANDLER_H
