+++ Tests/ConnectionHandler/Tests_ITableUsageByWorker.h
@@ -5,36 +5,29 @@
 #ifndef MOND_DB_TESTS_ITABLEUSAGEBYWORKER_H
 #define MOND_DB_TESTS_ITABLEUSAGEBYWORKER_H
 
-#include <gmock/gmock.h>
 #include "IWorker.h"
 #include "IHashTable.h"
 
-class MockTable: public IHashTable<std::vector<uint8_t>, std::vector<uint8_t>> {
-public:
-    ~MockTable() override = default;
-    MOCK_METHOD(bool, Insert, (const std::vector<uint8_t> &, const std::vector<uint8_t> &), (override));
-    MOCK_METHOD(bool, Remove, (const std::vector<uint8_t> &), (override));
-    MOCK_METHOD(bool, Find, (const std::vector<uint8_t> &), (const, override));
-    MOCK_METHOD(bool, Clear, (), (override));
-    MOCK_METHOD(size_t, GetSize, (), (const, override));
-    MOCK_METHOD(std::vector<uint8_t>, Get, (const std::vector<uint8_t> &key), (const, override));
-    MOCK_METHOD(bool, Update, (const std::vector<uint8_t> &, const std::vector<uint8_t> &), (override));
-    MOCK_METHOD(std::vector<std::vector<uint8_t>>, GetKeys, (), (override));
-private:
-    void grow(){}
-};
 
 class CheckingWorkersForITableUsage : public testing::Test{
+
+    void SetUp() override {
+        request._table_name = "MyFirstTable";
+        ON_CALL(storage, GetTable("MyFirstTable")).WillByDefault(Return(&table));
+    }
+
 public:
     Request request;
-    ITableStorage* storage = nullptr;
+    Request answer;
+    MockStorage storage;
     MockTable table;
     workerFactory factory;
 };
 
 TEST_F(CheckingWorkersForITableUsage, insertValueWorker) {
     request._type = requestType::INSERT;
-    auto worker = factory.get(request, storage);
+    auto worker = factory.get(request, &storage);
+    EXPECT_CALL(storage, GetTable(request._table_name)).Times(1);
     EXPECT_CALL(table, Insert(request._key, request._value)).Times(1);
     worker->operate();
     delete worker;
@@ -42,7 +35,8 @@ TEST_F(CheckingWorkersForITableUsage, insertValueWorker) {
 
 TEST_F(CheckingWorkersForITableUsage, removeWorker) {
     request._type = requestType::REMOVE;
-    auto worker = factory.get(request, storage);
+    auto worker = factory.get(request, &storage);
+    EXPECT_CALL(storage, GetTable(request._table_name)).Times(1);
     EXPECT_CALL(table, Remove(request._key)).Times(1);
     worker->operate();
     delete worker;
@@ -50,7 +44,8 @@ TEST_F(CheckingWorkersForITableUsage, removeWorker) {
 
 TEST_F(CheckingWorkersForITableUsage, clearWorker) {
     request._type = requestType::FIND;
-    auto worker = factory.get(request, storage);
+    auto worker = factory.get(request, &storage);
+    EXPECT_CALL(storage, GetTable(request._table_name)).Times(1);
     EXPECT_CALL(table, Find(request._key)).Times(1);
     worker->operate();
     delete worker;
@@ -58,7 +53,8 @@ TEST_F(CheckingWorkersForITableUsage, clearWorker) {
 
 TEST_F(CheckingWorkersForITableUsage, getSizeWorker) {
     request._type = requestType::GET_SIZE;
-    auto worker = factory.get(request, storage);
+    auto worker = factory.get(request, &storage);
+    EXPECT_CALL(storage, GetTable(request._table_name)).Times(1);
     EXPECT_CALL(table, GetSize).Times(1);
     worker->operate();
     delete worker;
@@ -66,7 +62,8 @@ TEST_F(CheckingWorkersForITableUsage, getSizeWorker) {
 
 TEST_F(CheckingWorkersForITableUsage, getWorker) {
     request._type = requestType::GET;
-    auto worker = factory.get(request, storage);
+    auto worker = factory.get(request, &storage);
+    EXPECT_CALL(storage, GetTable(request._table_name)).Times(1);
     EXPECT_CALL(table, Get(request._key)).Times(1);
     worker->operate();
     delete worker;
@@ -74,7 +71,8 @@ TEST_F(CheckingWorkersForITableUsage, getWorker) {
 
 TEST_F(CheckingWorkersForITableUsage, updateWorker) {
     request._type = requestType::UPDATE;
-    auto worker = factory.get(request, storage);
+    auto worker = factory.get(request, &storage);
+    EXPECT_CALL(storage, GetTable(request._table_name)).Times(1);
     EXPECT_CALL(table, Update(request._key, request._value)).Times(1);
     worker->operate();
     delete worker;
