+++ Tests/ConnectionHandler/tests.cpp
@@ -1,13 +1,10 @@
 // Copyright 2022 MOND
 
 #include <gtest/gtest.h>
-#include "Tests_ConnectionHandler.h"
 #include "Tests_IStorageUsageByWorker.h"
 #include "Tests_ITableUsageByWorker.h"
 #include "Tests_WorkerFactory.h"
 
-
-TEST(Test, test_for_test) { EXPECT_TRUE(true); }
 
 int main() {
     ::testing::InitGoogleTest();
